Merge branch 'feature/12-refactor-and-add-core-ds-charhashset' into 'draft/develop'
Data-structures to be added / Refactor and add core data structures to core package rather than strhelper or other

Closes #12

## Data-structures to be added / Refactor and add core data structures to core package rather than strhelper or other

- JSON parsing added with collections and data models
- Hashset
- HashsetsCollection
- CharHashsetMap

## Background why needed?

Previously common data structures were added to the `strhelper` package, however, it seems like those structures should be available in the core or root level than the specific package.

### Problem Encounter

We want the `core` package to be the root of all packages and `strhelper` is also a core package, such core packages like `strhelper`, `pathhelper` should not depend on each other but all depend on the `core` package, and in specific cases `pathhelper` requires data structures which currently resides in `strhelper`.

### Solution

We need to include the core data structures to reside in the core rather than any other package.

## Worked on Items

- [x] #12 - Hashset
- [x] #12 - HashsetsCollection
- [x] #12 - CharHashsetMap

## TODOs

NA

## References

- [Input Codes](https://pastebin.ubuntu.com/p/dzdYPXNKjk/)
- [Output Logs](https://pastebin.ubuntu.com/p/5RfsbCBt7b/)
- [go - Calling a method on a nil struct pointer doesn't panic. Why not? - Stack Overflow](https://stackoverflow.com/questions/42238624/calling-a-method-on-a-nil-struct-pointer-doesnt-panic-why-not)
- [Array of pointers to JSON - Stack Overflow](https://stackoverflow.com/questions/28101966/array-of-pointers-to-json)
- [Json Parsing of Array Pointers](https://play.golang.org/p/zTuMLBgGWk)

## Code Samples

- [Input Codes](https://pastebin.ubuntu.com/p/dzdYPXNKjk/)
- [Output Logs](https://pastebin.ubuntu.com/p/5RfsbCBt7b/)

See merge request evatix-go/core!16